+++ backend/src/app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Numeric
+from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
 
 from app.core.database import Base
@@ -12,6 +12,6 @@ class User(Base):
     last_name = Column(String, nullable=False)
     email = Column(String, unique=True, index=True, nullable=False)
     password = Column(String, nullable=False)
-    experience_points = Column(Numeric, default=0)
+    experience_points = Column(Integer, default=0)
     user_type = Column(String, nullable=False)
     gender = Column(String, nullable=True)
